use long syntax for other types as well (tests not working)
